Simplify the code in Shell.tsx and move sequence utils to a utils file
